checked CRUD
Checked CRUD ops again for Housing, Discr_exp. Added remove all for users (Per Brandon) All good

+++ code/src/app/models/discr_expense.model.js
@@ -1,7 +1,7 @@
 // Josiah Stadler 3/16
 const sql = require("./db.js");
 
-const  Discr_expense = function(discr_expense) {
+const  Discr_expense = function(_discr_expense) {
   this.DISC_EXPID = _discr_expense.DISC_EXPID;
   this.NAME = _discr_expense.NAME;
   this.AMOUNT = _discr_expense.AMOUNT;
@@ -11,7 +11,7 @@ const  Discr_expense = function(discr_expense) {
 };
 
 Discr_expense.create = (_discr_expense, result) => {
-  sql.query("INSERT INTO discr_expense(NAME, AMOUNT, HOW_OFTEN, PRIOTITY, USERID) VALUES(?, ?, ?, ?);",[_discr_expense.NAME, _discr_expense.AMOUNT, _discr_expense.HOW_OFTEN, _discr_expense.PRIORITY, _discr_expense.USERID], (err, res) => {
+  sql.query("INSERT INTO discr_expense(NAME, AMOUNT, HOW_OFTEN, PRIORITY, USERID) VALUES(?, ?, ?, ?, ?);",[_discr_expense.NAME, _discr_expense.AMOUNT, _discr_expense.HOW_OFTEN, _discr_expense.PRIORITY, _discr_expense.USERID], (err, res) => {
     if (err) {
       console.log("error: ", err);
       result(err, null);
@@ -38,7 +38,7 @@ Discr_expense.getAll = result => {
   });
 };
 Discr_expense.getUserDiscr_expense = (_discr_expense, result) => {
-  sql.query("SELECT * FROM debt WHERE USERID = ?;",[_discr_expense.USERID], (err, res) => {
+  sql.query("SELECT * FROM discr_expense WHERE USERID = ?;",[_discr_expense.USERID], (err, res) => {
     if (err) {
       console.log("error: ", err);
       result(null, err);
@@ -49,10 +49,10 @@ Discr_expense.getUserDiscr_expense = (_discr_expense, result) => {
     result(null, {...res });
   });
 };
-Discr_expense.updateDiscrExp = (DISC_EXPID, USERID, result) => {
+Discr_expense.updateDiscrExp = (_discr_expense, result) => {
   sql.query(
     "UPDATE discr_expense SET NAME = ?, AMOUNT = ?, HOW_OFTEN = ?, PRIORITY = ? WHERE DISC_EXPID = ? and USERID = ?",
-    [_discr_expense.NAME, _discr_expense.AMOUNT, _discr_expense.HOW_OFTEN, _discr_expense.PRIORITY,  DISC_EXPID, USERID],
+    [_discr_expense.NAME, _discr_expense.AMOUNT, _discr_expense.HOW_OFTEN, _discr_expense.PRIORITY,  _discr_expense.DISC_EXPID, _discr_expense.USERID],
     (err, res) => {
       if (err) {
         console.log("error: ", err);
@@ -66,14 +66,28 @@ Discr_expense.updateDiscrExp = (DISC_EXPID, USERID, result) => {
         return;
       }
 
-      console.log("updated discrectionary expense with ID: ", { DISC_EXPID: DISC_EXPID, ..._discr_expense });
-      result(null, { DISC_EXPID: DISC_EXPID, ..._discr_expense });
+      console.log("updated discrectionary expense with ID: " + _discr_expense.DISC_EXPID);
+      result(null,res);
     }
   );
 };
-
-Discr_expense.remove = (DISC_EXPID, result) => {
-  sql.query("DELETE FROM discr_expense WHERE DISC_EXPID = ?", DISC_EXPID, (err, res) => {
+Discr_expense.findByDiscrExpId = (_discr_expense, result) => {
+  sql.query("SELECT * FROM discr_expense WHERE DISC_EXPID = ?" , [_discr_expense.DISC_EXPID], (err, res) => {
+    if (err) {
+      console.log("Sql Error!");
+      result(null, err);
+      return;
+    } if (res.length == 0) {
+      console.log("No Discretionary Expense found by the ID: " + _discr_expense.DISC_EXPID);
+      result(err);
+      return;
+    }
+    console.log("Discretionary Expense ID found for:" + _discr_expense.DISC_EXPID)
+    result(null, res);
+  });
+};
+Discr_expense.remove = (_discr_expense, result) => {
+  sql.query("DELETE FROM discr_expense WHERE DISC_EXPID = ?",_discr_expense.DISC_EXPID, (err, res) => {
     if (err) {
       console.log("error: ", err);
       result(null, err);
@@ -86,23 +100,29 @@ Discr_expense.remove = (DISC_EXPID, result) => {
       return;
     }
 
-    console.log("deleted discrentionary expense with ID: ", DISC_EXPID);
+    console.log("deleted discrentionary expense with ID: ",  _discr_expense.DISC_EXPID);
     result(null, res);
   });
 };
 
-Discr_expense.removeAll = result => {
-  sql.query("DELETE FROM discr_expense", (err, res) => {
-    if (err) {
-      console.log("error: ", err);
-      result(null, err);
-      return;
-    }
+Discr_expense.removeAll = (_discr_expense, result) => {
+  sql.query("DELETE FROM discr_expense WHERE USERID = ?", [_discr_expense.USERID], (err, res) =>{
+      if (err) {
+          console.log("error: ", err);
+          result(null, err);
+          return;
+      }
 
-    console.log(`deleted ${res.affectedRows} discretionary expense entries`);
-    result(null, res);
+      if (res.affectedRow == 0) {
+          result("Not Found:" + null);
+          return;
+      }
+
+      console.log("Deleted ALL Discretionary Expenses for user with ID: ", _discr_expense.USERID);
+      result(null, res);
   });
 };
+
 
 
 module.exports = Discr_expense;
